requirements.txt
langchain==0.3.25
langchain-openai==0.3.22
langchain-community==0.3.25
chromadb==1.0.12
pyyaml==6.0.2
python-dotenv==1.1.0
streamlit==1.32.2
